change timePeriod from double to int, add chart

+++ lib/features/analysis/domain/services/analysis_service.dart
@@ -12,7 +12,7 @@ class AnalysisService {
 
   AnalysisResult getAverageSpeedForPeriod(
     List<SpeedTime> data,
-    double timePeriod,
+    int timePeriod,
   ) {
     if (data.isEmpty) {
       return AnalysisResult(
